fix type error bug

+++ model.py
@@ -181,7 +181,7 @@ class HumanPartsNet(chainer.Chain):
         mask0 = predictions.argmax(axis=1).reshape(mask1.shape) > 0
         intersection = xp.bitwise_and(mask0, mask1).sum(axis=1)
         union = xp.bitwise_or(mask0, mask1).sum(axis=1)
-        return (intersection / union).mean(dtype = predictions.dtype)
+        return (intersection.astype(predictions.dtype)  / union.astype(predictions.dtype)).mean()
 
 
 def load_VGGmodel():
